+++ src/commands/rank.ts
@@ -1,5 +1,5 @@
 import Command from '../interfaces/command';
-import { Message, Attachment, User, GuildMember } from 'discord.js';
+import { Message, Attachment, GuildMember } from 'discord.js';
 import bot from '..';
 import { createCanvas, loadImage, Image, Canvas } from 'canvas';
 import { User as DatabaseUser } from '../interfaces/databaseStructures';
@@ -45,7 +45,12 @@ export default class RankCommand implements Command {
         ctx.drawImage(image, 40, 30, 160, 160);
 
         ctx.font = '50px Fredoka One';
-        ctx.fillText(member.displayName, 730 - ctx.measureText(member.displayName).width, 65);
+        let size = ctx.measureText(member.displayName).width;
+        if(size > 500)
+            size = 500;
+
+        console.log(size);
+        ctx.fillText(member.displayName, 730 - size, 65, 500);
 
         ctx.fillRect(230, 150, ((data.messages - 200 * data.level) / 200) * 525, 40);