+++ src/interfaces/databaseStructures.ts
@@ -1,6 +1,7 @@
 export interface Permission {
     action: string;
     user_ids: string[];
+    _id?: string;
 }
 
 export interface User {
@@ -8,16 +9,21 @@ export interface User {
     level: number;
     messages: number;
     guild: string;
+    cooldown: number;
+    _id?: string;
 }
 
 export interface Server {
     id: string;
     language: string;
-    ranking?: boolean;
+    ranking: boolean;
+    prefix: string;
+    _id?: string;
 }
 
 export interface Tag {
     server: string;
     title: string;
     content: string;
+    _id?: string;
 }