+++ webpack.config.prod.js
@@ -1,9 +1,6 @@
-const { VueLoaderPlugin } = require("vue-loader");
-const CopyWebpackPlugin = require("copy-webpack-plugin");
-
-const webpack = require('webpack')
-
-const { resolve } = require("path");
+const { VueLoaderPlugin } = require('vue-loader');
+const CopyWebpackPlugin = require('copy-webpack-plugin');
+const { resolve } = require('path');
 
 module.exports = {
     mode: 'production',
@@ -40,15 +37,6 @@ module.exports = {
         new CopyWebpackPlugin([{
             from: './src/dashboard/public/index.html',
             to: '.'
-        }]),
-
-        /*
-        new HtmlWebpackPlugin({
-            title: 'izel',
-            template: './webpack-template.html',
-            inject: 'body',
-            alwaysWriteToDisk: true
-        }),
-        */
+        }])
     ],
-}
+};